requirements.txt
absl-py==1.2.0
dm-haiku==0.0.7
jaxline==0.0.5
ml_collections==0.1.1
numpy==1.23.2
optax==0.1.3
tensorflow==2.9.1
tensorflow_datasets==4.6.0
